Use find_allowed in locate if model_class exists
When using a custom locator to implmenent a non-standard "find",
`gid.model_class` could still trigger an exception if the class does not
exist.

This simply checks if the `model_name` constant is defined before
running the `find_allowed?` check.

In the context of ActiveJob, a deserialization error would still occur
in the default case, but it would be triggered by `locator#locate`
instead of in this method.